tpm2_identity_util: replace "encrypt_seed" with "share_secret"
This commit changes the
tpm2_identity_util_encrypt_seed_with_public_key in two ways:

1. It changes the callee to always be responsible for generating the
seed, independent of the public key algorithm. Previously, the caller
had to generate a random seed for RSA keys, which was unnecessary for
ECC keys.

2. It renames "encrypt_seed" to "share_secret" to be more consistent
with the name that the TPM 2.0 Library Specification uses for this
mechanism.

Fixes #1730.

Signed-off-by: Matthew Dempsky <matthew@dempsky.org>