lib/openssl: pad ECC key params
When reading an ECC key from a file, one of it's parameters (private
scalar, public coordinates x and y) can start with leading zero bytes.
The BN_num_bytes() returns the bytes required to store such a parameter,
i.e. not the key size if the parameter starts with 0x00.
Additionally, BN_bn2bin does not pad the parameter to its expected size
which can lead to problems in the TPM.
Hence, the key sizes are calculated properly and ECC parameters are padded.

Signed-off-by: Johannes Holland <johannes.holland@infineon.com>