tpm2_nvdefine: add support for friendly NT names
When I first wrote the attribute code, it wasn't clear to me
what the NT field was for, which is specifying the type of the
NV index, such as counter or bits. Add support for using
friendly names, like counter or bits. Also, correct the
manpage which incorrectly showed setting an NT field of 3,
which is not a valid selector.

In the future, we could probably drop the nt= portion and just
specify the NT names directly in the string option and validate
that no one specifies both ways or conflicting choices.

Signed-off-by: William Roberts <william.c.roberts@intel.com>