tpm2_evictcontrol: fix flushing logic
tpm2_evitccontrol uses the handle of the loaded object to determine if it should
evict or persist when an argument is not given. In the case of loading from an
ESYS_TR serialized handle, their is no way to determine if the object is persistent
or transient so we can't figure out what ouytput to provide.

This makes partial progress on #1816 and at least indicates if it is
evicted or not.

Signed-off-by: William Roberts <william.c.roberts@intel.com>