test: add is_alg_supported
Add a bash function is_alg_supported to helpers.sh. This function uses
tpm2_testparams to determine if an algorithm is supported. This does
not work for hashes, however, there are already utility functions
implemented in helpers.sh.
The function returns 0 if the algorithm is supported and 0 otherwise.
If the TPM is a simulator, the function will error if an algorithms
is unsupported. This prevents errors in the simultator causing untested
code.

Signed-off-by: Johannes Holland <johannes.holland@infineon.com>