test: add is_cmd_supported and replace encryptdecrypt
Some commands (EncryptDecrypt in particular) are not supported on all
TPMs. We can only assume commands specified as mandatory in the PCClient
spec to be implemented. However, we want to test the full functionality
of the tools.
Thus, a function is_cmd_supported is added to helpers.sh which returns 1
if a command is unsupported by the TPM. At the same time, if the TPM is
a simulator, this function will error if a command is unsupported.
EncryptDecrypt was replaced for TPMs without support for this command:
 - for asymmetric keys by sign/verify
 - for symmetric keys by readpublic
Testing of NV_GlobalWriteLock will be omitted for non-simulator TPMs
without support for this command.

Signed-off-by: Johannes Holland <johannes.holland@infineon.com>