tpm2_eventlog: Add tool to parse and display a TPM2 event log as yaml.
The format for the event log parsed by this tool is documented in
section 9 of the "TCG PC Client Platform Firmware Profile Specification
(aka "PFP"). This is the "crypto-agile" format capable of representing
digests of varying size. Some constants / values necessary for
understanding the log contents are spread out over other specs but the
structure is documented in the PFP.

This log is a series of concatinated packed binary structures describing
events generated / collected by the firmware on platform boot. Each
'event' has 3 primary components: First is a header describing the PCR
extended, the 'type' of the event and the number of digests in the event.
After this a series of digest structures follows. The final data in this
structure is the event size and a binary representation of the event. The
total size of the log must be inferred.

This tool takes the log file as it's only parameter (positional). It
parses the log and writes a yaml / text representation to stdout. The
option processing and 'main' function for the tool is located in
'tools/misc/tpm2_eventlog.c'. If the log file parameter is omitted the
tool will return an error.

The log parsing logic and the text formatting logic are located in the
'lib/tpm2_eventlog' and 'lib/tpm2_eventlog_yaml' modules respectively.
The main driver initiates the log parsing through the 'yaml_eventlog'
function that takes a reference to the start of the log and its size.
yaml module then iterates over the eventlog using the 'foreach_event2'
function from the eventlog parsing module. This function is provided
with and a callback function that displays info about the event in yaml.
The parser logic validates the event structure before invoking
callbacks.

Additionally this commit adds unit tests and a single integration test.
Unit tests are used to test the parser and display logic. The integration
test invokes the tool passing it a single binary event structure taken
from a VM booting the OVMF firmware with an emulated TPM2 backend.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>