tpm2_eventlog: Parse and display UEFI variable events.
EDK2 firmware records the value of a numnber of variables (primarily
related to secureboot) in the event log. This commit updates the
tpm2_eventlog module so that it validates the structure of these events.
The tpm2_eventlog_yaml module is updated as well to display them. These
events contain GUIDs and so this commit adds a dependency on libuuid to
simplify parsing.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>