tpm2_eventlog: Break event processing into 3 callbacks.
This commit refactors the tpm2_eventlog module to make processing of
individual events more straight forward. Instead of requiring consumers
to manually trigger the processing of digests themselves we consolidate
all of the logic in the tpm2_eventlog module.

The foreach_event2 function now takes 3 callbacks: The first is the
original callback passed the TCG_EVENT_HEADER2 structure. The second is
invoked for each TCG_DIGEST2 in the event. The third is passed the
TCG_EVENT2DATA from the event. All of the callbacks are invoked in the
order that the data appears in the event log.

Any, or all of these callbacks can be omitted. In this case the
tpm2_eventlog module will validate the structure of the eventlog
returning the appropriate value.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>