tpm2_eventlog: Parse and display UEFI action events.
EDK2 firmware records a number of strings into the event log before
particular actions are carried out. These are simple strings (not NULL
terminated though) that can be displayed to the caller. There is no
structure to these events beyond the string and so only the
tpm2_eventlog_yaml module requires changes. An integration test for this
particular event is added as well.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>