tpm2_eventlog: Parse and display the SpecID event.
The platform firmware profile requires firmware to make the first event
in the log a descriptive event. This event is called the SpecID event
and it documents the hashes supported by the platform as well as a
number of other attributes describing the firmwares spec compliance
(version etc).

The SpecID event does not use the TPM2 event format. Instead it uses the
older format from the 1.2 TPM to allow version / format discovery by
parsers that only understand the 1.2 format. This makes the
'foreach_event2' interface incapable of parsing the full log. This
commit adds a new function to the tpm2_eventlog module 'parse_eventlog'
that takes a new callback specific to the SpecID event.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>