tpm2_eventlog: Parse and display UEFI services events.
Entries in the log relevant to UEFI services format their event using
the UEFI_IMAGE_LOAD_EVENT structure. This commit adds code to validate
this structure in the tpm2_eventlog module and code to display it in the
tpm2_eventlog_yaml module.

NOTE: The 'DevicePath' member of the UEFI_IMAGE_LOAD_EVENT structure is
treated as a byte array and dispayed as a hex string. Per the spec it
should be interpreted as a UEFI_DEVICE_PATH or one of its decendents. We
do this for two reasons: First these structures are pretty complicated
and would require pulling in UEFI headers. Second the OVMF firmware from
fedora 31 produces what look to be malformed structures (all data 0'd)
which prevents us from knowing how to interpret the device path.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>