tpm2_util: fix public to yaml exponent handling
The routine converting TPM2B_PUBLIC to YAML was just outputting the
exponent as is, and in hex. The TPM uses 0 to indicate the default
exponent which 65537. The YAML would just output 0x0. So add a condition
where if 0, output 65537, and update the output to unsigned singe that
is commonly how exponents are reported.

Before:
type:
  value: rsa
  raw: 0x1
exponent: 0x0
bits: 2048

After:
type:
  value: rsa
  raw: 0x1
exponent: 65537
bits: 2048

Fixes: #1919

Signed-off-by: William Roberts <william.c.roberts@intel.com>