lib/files: don't rely on the reported size in file_read_bytes_from_file
The reported file size is not always correct, e.g. for sysfs files
generated on the fly by the kernel when they are read, which appear as
having size 0. Read as many bytes as we can until EOF is reached or the
provided buffer is full. As a small sanity check, fail if the number of
bytes read is smaller than the reported file size.

This change leads to a problem with the current approach to wrapping
fread in test_tpm2_auth_util: it unconditionally memcpy's the requested
number of bytes regardless of whether there is actually enough mocked_file_data
for that. This leads to a test failure when memory sanitation is
enabled. Make sure not to copy unitialised memory by using strncpy
instead of memcpy.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>