tools/misc/tpm2_eventlog: don't rely on the reported file size
When using tpm2_eventlog directly on the sysfs file
/sys/kernel/security/tpm0/binary_bios_measurements, an empty buffer is
allocated because its file size is reported as 0. As we cannot know the
correct buffer size in advance, allocate a buffer for the maximum amount
of data and use the ability of file_read_bytes_from_file from the
previous commit to determine the correct buffer size from a full file
read.

Fixes: GH-1975
Signed-off-by: Jonas Witschel <diabonas@gmx.de>