[RFC] tpm2_netlisten: introduce new command
This command is designed to use the network TCTI to listen and accept
client connections on a plain TCP socket. It will than forward the
command buffers to the TCTI of choice (selected by --tcti option) and
return response buffer to the client. NOTE that this is an unsecured,
plain text socket, so securing with a tunnel or VPN is encouraged.

TODO:
- Complete manpage
- Perhaps add a startup warning to tunnel unless -y/--yes option is
  specified.

Signed-off-by: William Roberts <william.c.roberts@intel.com>