tpm2_createak: add qualified name output
Add the qualified name output of the AK. The qualified name of
an object is the name hash alg of the qualified name of the parent and
the name of the object. Thus the qualified name serves as a hash of the
loaded object chain from the hiearchy seed to the object.

To quote the spec:
The Qualified Name (QN) of an object is the digest of all of the Names of all of the ancestor keys back to
the Primary Seed at the root of the hierarchy. The QN of an object includes the Name of the object. The
QN uses the Name hash of the current object to compute the QN for the object.

Signed-off-by: William Roberts <william.c.roberts@intel.com>