tpm2_readpublic: add qualified name to file output
Add the qualified name as a file output to tpm2_readpublic
via the -q option.

The qualified name. TO qoute the spec:
The Qualified Name (QN) of an object is the digest of all of the Names of all of the ancestor keys back to
the Primary Seed at the root of the hierarchy. The QN of an object includes the Name of the object. The
QN uses the Name hash of the current object to compute the QN for the object.

Assuming that key A is the Storage Parent of object B, then the Qualified Name of B (QNB) is:
QNB ≔ HB (QNA || NAMEB)

Signed-off-by: William Roberts <william.c.roberts@intel.com>