tss2_*: tss2_setappdata now reads from file or stdin
The command tss2_setappdata falsely allowed only a parameter argument
to be written, which disables the ability to write binary data. This
is fixed by switching the parameter to the file/stdin pattern used
elsewhere in order to enable binary data writing. Manpages, integration
tests and autocompletion are adapted accordingly.

Signed-off-by: Christian Plappert <christian.plappert@sit.fraunhofer.de>