configure.ac: make -W cflag checks more robust
Stumbled upon this while attempting to compile the codebase
(unsuccessfully) with clang. As it happens, the underlying autoconf
macro (AX_CHECK_COMPILE_FLAG) will only fail if the compiler returns
non-zero. After a few moments of exploration, this is not always
sufficient.

In the case of clang, an unrecognized cflag will emit a warning, but
clang will return 0. Therefore, configure will (incorrectly) conclude
the cflag is recognized. See #2097 for the result.

In the case of gcc, if a cflag is recognized but used incorrectly
(example: -Wformat-security by itself), gcc will emit a warning but
return 0.

Therefore, if -Werror is turned on at build time, and an unrecognized
cflag is encountered, configure will succeed, but the build will fail.

So, add smarts (EXTRA-FLAGS) to the macros, which results in clang or
gcc exiting with an error should it encounter a cflag it does not
recognize (or a cflag used incorrectly).

These smarts revealed, with gcc, that -Wformat-security cannot be
checked on its own, so correct how we check it. It also revealed that
-Wbool-compare is not a valid cflag for clang, so move that check into
the hardening flags block.

Signed-off-by: Joe Konno <joe.konno@intel.com>