tool: use atexit() for cleanup
Register the exit routines with atexit(), so calls to exit() will invoke
cleanup routines. Also, use a cleanup routine for the main program. This
way, tools can call exit when they need and the main program as well as
the tool will have calls for any cleanup they need to perform.

Signed-off-by: William Roberts <william.c.roberts@intel.com>