eventlog: Add handling of sha1 log format (Fixes #2179).
If no specid event is part of the event list no description of the crytpo agile event
format is available. In this case it will be assumed that the sha1 log format is
used.
A function for parsing the sha1 log event header and yaml callbacks for this format
were added.
A fixture with a binary sha1 event log was added to the integration tests.

Signed-off-by: Juergen Repp <juergen.repp@sit.fraunhofer.de>