coverity: only perform one publish
Currently all GCC builds for each docker container (DOCKER_TAG) are
being submitted to coverity, for example see build:
  - https://travis-ci.org/github/tpm2-software/tpm2-tools/builds/719995881

To prevent this, set a COVERITY_DOCKER_TAG variable in Travis and make
this available to the coverity_scan branch. Ensure that DOCKER_TAG is
equal to COVERITY_DOCKER_TAG to filter out builds to a specific docker
image. This will prevent multiple submissions to coverity.

Signed-off-by: William Roberts <william.c.roberts@intel.com>