man_to_bashcompletion: Add utility to create bash completion scripts
The utility takes iterates through the man pages in the man directory
to generate bash completion scripts for all tools. This utility
benefits from the fact that the man pages are already checked against
the tools directory and so should be consistent with tool updates.

The bash completion scripts generated by this utility can be run with
eval "`/path/to/this/utility/man_to_bashcompletion.sh`"

Signed-off-by: Imran Desai <imran.desai@intel.com>