Fix tpm2_checkquote when run on a big endian system
Various PCR related structures read by tpm2_checkquote are written
in little endian format, which needs to be taken into account when
run on a big endian system. Although little endian is unusual for
TPM, it is better to leave these data structures in little endian
format to not disrupt existing systems.

This patch creates new versions of those functions where the input
format is in little endian when it is read from files. The existing
version of the functions handles the case when the data is written
in native format, so no conversion is needed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>