tss2 *: The tss2 * tools now support to use TPM objects protected by a policy with PolicySigned
In case the TPM object (e.g., key or NV index) is protected by a policy that has a
PolicySigned command, the user is now prompted to sign a TPM-provided challenge
with the required authentication key.

This was done by implementing a callback function for the command
Fapi_SetSignCB and create the corresponding integration test.

Signed-off-by: Christian Plappert <christian.plappert@sit.fraunhofer.de>