lib/tpm2.c: phash alg/size undeterministic for ESYS_TR_PASSWORD
Affix phash alg/size to TPM2_ALG_SHA256 for the tools that have
not yet moved to the new template that parses session type and
session attribute information to calculate phash information.

This is required as a workaround for turning off hmac sessions
on empty auths which causes the cphash calculations to fail.

Signed-off-by: Imran Desai <imran.desai@intel.com>