tpm2_eventlog: Print partition name correctly
Currently with --eventlog-version=2, only the first character of a
partition name will be printed, as the parser interprets it as UTF8. Per
the spec, these are UTF16 (similar to the UnicodeName of an EFI
variable).

We move all such strings to use UTF16_CHAR which is essentially a PACKED
char16_t, so that they will be handled correctly.

The public function yaml_uefi_var_unicodename() is replaced with the
private function yaml_utf16_to_str() so we can use the same code in both
places where we need to print out a UTF16 string.

Signed-off-by: Joe Richey <joerichey@google.com>