man: document non-supported RM workaround in startauthsession
When an RM is encountered that doesn't perform session ungapping, a
workaround exists to start a longer running seperate process that
communicates with the RM, and send the commands to it. This way the
RM won't flush after each command invocation, as the RM only sees the
longer running process. This will require the user to flush transient
handles until tpm2-tools starts flushing these things, see #1511.

Signed-off-by: William Roberts <william.c.roberts@intel.com>