policy: Support TPMLess commands
For now this works with policysecret, owner hiearchy and NULL auth.
It's hardcoded to ignore the TPM.

tpm2 policysecret -S session.ctx -c o
0d84f55daf6e43ac97966e62c9bb989d3397777d25c5f749868055d65394f952

TODO:
For each policy command:
  - Support --tcti=none
  - When --tcti is none, require:
    --name/-n
    -L/--policy for old value
    -- Require hash algorithm.
    - Don't output a TICKET, since we can't?

Todo, consider creating a session.ctx structure that can be passed
from startauthsession with --tcti=none? This was we can encapsulate
the state instead of requiring -L and hash algorithm?

Signed-off-by: William Roberts <william.c.roberts@intel.com>