lib/tpm2_nv_util.h: parse info for NV op specific max nv buffer sz
The NV index size is constrained to different data sets read from
the GetCapability structures because there is a dependency on the
NV operation type: Define vs Read vs Write vs Extend.

This commit addresses those parameters and chooses a sane default
in the case where GetCapability fails or fails to report the
specific property/ data set. This is especially true because some
of the properties are TPM implementation dependent.

Signed-off-by: Imran Desai <imran.desai@intel.com>