hmac: update support to be consistent with create
Their was a discrepency between tpm2_create and tpm2_import with respect
to -G hmac. In the tpm2_create tool, it would generate a keyedhash
object with a scheme of hmac->sha256, however in import it would create
a keyedhash object with a NULL scheme. Support these in import but keep
them distinct and consistent with other tooling. Also, while at it, this
changes the structure of the openssl import routines to take the
template of the object it's creating along with the metadata. This way a
flow would be to populate a template using tpm2b_public_init and then
pass that template along to the openssl routines to generate a public
and sensitive portion off that template with the additional key
material.

Signed-off-by: William Roberts <william.c.roberts@intel.com>