[RFC] tools: add tpm2_policy tool for invoking libpolicy
UPDATE:
Working:
Initial test working:
If you get the tpm2_policy tool from the tss pr, this should work:
tpm2 policy /path/to/tpm2-tss/test/data/fapi/policy/pol_pcr16_0.json

Their are still many todo's to go through, but let's figure out the
interface into the library and go from there.

Create a tpm2_policy tool that can read the FAPI JSON style policies
and:

1. Instantiate them -> This process fills in anything missing in the
template. TODO: How does this get handled, do we need to tweak any
of the callbacks?

2. Calculate them -> This process produces a list of hashes... TODO:
   Why? Is this a list of all the subordinate policies or can the json file
   have N policies where N > 1?

3. Execute them -> Execute the policy on a session. TODO: Who is
supposed to start the policy session as it seems to be 0?

Their are a lot of TODO items in this code. I'm looking for how we want
to use this, different tools, like tpm2_policyinit tpm2_policycalc and
tpm2_policyexec? Or an all in-one tool. Currently note that ONLY the
Execute needs an ESYS context, but instantiate should be filling stuff
in so it likely needs a context or the callbacks handled?

Signed-off-by: William Roberts <william.c.roberts@intel.com>