tools/tpm2_nvread: fix reads > TPM2_PT_NV_BUFFER_MAX with policy session
Fixes #2875

Policy authorization session must be re-instantiated on multiple
iterations of nv read if the read size exceeds the allowed operating
size. To avoid this, only a single iteration is allowed when using
policy authorization session.

Re-instantiation is avoided because compounded policies are not
possible to be replayed to arrive at the same policy digest.

Signed-off-by: Imran Desai <imran.desai@intel.com>