tpm2_createak: use Esys_TR_GetName over ReadPublic
Rather than going to the TPM to get the name alg for the EK via a
TPM2_ReadPublic call, just use the name as cached by ESYS. The name of
an object is the name algorithm followed by the hash itself.

Not only is this faster, but prevents a MITM from lying to use about the
EK properties as the TPM2_Readpublic was not protected with a session.

Signed-off-by: William Roberts <william.c.roberts@intel.com>