fapi: fix usage of parameter pcrLog for tss2_quote.
pcrLog is an optional parameter. If pcrLog is not used as parameter currently
the pcr log is still calculated in Fapi_Quote.
To avoid this calculation a NULL pointer will be passed to Fapi_Quote if the
parameter pcrLog is not passed.
So tss2_quote can be executed for a user which has no access rights to the
files with the system measurements.

Signed-off-by: Juergen Repp <juergen.repp@sit.fraunhofer.de>