tpm2_eventlog: fix read loop for files bigger than one chunk
The read loop is not correctly accumulating the total file size which
will result in erroneous behaviour should a file larger than one chunk
be encountered.

While fixing this also remove redundancy in the allocation logic and
avoid an extraneous chunk allocation even if the EOF has already been
reached.

Signed-off-by: Matthias Gerstner <matthias.gerstner@suse.de>