test: add missing shut_down call on cleanup
The tests "gettime", "readclock" and "symlink" are not calling
"shut_down" during the "cleanup" stage, making the TPM simulator to keep
their process during the "make check".  Somehow this produces problems
when the tests are executed in parallel under certain conditions, with
the effect of "make" not ending and waiting to those process to die.

This commit and the mising call in the "cleanup" stage.

Fix #3042

Signed-off-by: Alberto Planas <aplanas@suse.com>