tools/tpm2_certify: fix man page for short options and add tests
Fixes #3082

The short options for the signing-key-auth and certified-key-auth
were swapped. The case fix in the man page makes it less intuitive
but have to go through with the change so we don't break any
existing scripts.

This change does not affect the long options. Tests have been
added to ensure the functionality.

Signed-off-by: Imran Desai <imran.desai@intel.com>