tpm2_eventlog_yaml: fix malformed YAML for EV_IPL data
The code for printing EV_IPL data was fairly crude and often
did not generate valid YAML syntax. Some problems

 * Data starting with a space would result in invalid
   indentation, a leading space requires a quoted string
 * Non-printable cahracters must generally be escaped,
   using a quoted string
 * Embedded NUL bytes were turned into newlines, which
   mangled any UTF16 encoded data.

This change attempts to make the YAML output much safer. It
is not pefect as it just processes the data bytewise and
thus could potentially emit invalid UTF-8 bytes. In practice
this won't be a problem for known bootloader emitting EV_IPL
events.

This changes the formatting slightly

  - All strings are now surrounded with double quotes

  - All NUL bytes, including the final trailing NUL
    are displayed in escaped format.

  - Non-printable ASCII chars are escaped, including
    the tab character, per YAML recommendations

A much better long term solution would be to switch to
using libyaml for generating the output which would give
a strong guarantee of correct formatting.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>