tpm2_encodeobject: fix auth boolean flag
The flag for wether or not a key needs a password was being set based on
if the parent needed a password or not when it should be set based on if
the child object needs a password or not.

Correct this by adding a -p/--key-auth option to indicate the value of
this boolean.

$ tpm2 encodeobject -C 0x81000000 -u key.pub -r key.priv -o key.pem
$ openssl asn1parse -dump -inform PEM -in key.pem
   <snip>
   14:d=2  hl=2 l=   1 prim: BOOLEAN           :0
   </snip>

$ tpm2 encodeobject -C 0x81000000 -u key.pub -r key.priv -o key.pem -p
$ openssl asn1parse -dump -inform PEM -in key.pem
   <snip>
   14:d=2  hl=2 l=   1 prim: BOOLEAN           :1
   </snip>

A workaround would be manually modifying the ASN1 PEM file boolean flag
OR creating the same parent key but with a password and specifying the
password via `-P`. Note that a primary key is the same given the same
inputs and password doesn't change the generated key.

Fixes: #3152

Signed-off-by: William Roberts <william.c.roberts@intel.com>