tpm2_startauthsession: VERIFY persistent handles
When an AuthSession is used with a tpmKey, that key MUST be verified in
order to trust the connection. Think of this as analagous to known hosts
with ssh. With Serialized ESYS_TR's or Context Files, the name is
associated with that handle and ESAPI is instrumented to perform all the
required name checks. When loaded from a persistent handle, ESAPI does
not have the associated metadata and reads the name from the handle with
a ReadPublic call. Becuase of this, an attacker COULD MiTM the
connection with their own key. Thus, with persistent handles we must
verify the name from some other source. The TPM name could be thought of
as an ssh key fingerprint.

To fix this, provide a -n argument for a name file as output by
readpublic or could be generated as a binary array of the objects name.
When a tpmKey is provided, check the name and fail.

Fixes: #2812

Signed-off-by: William Roberts <william.c.roberts@intel.com>