policy authorization: allow the file attribute for policy authorization.
Currently for several commands it's necessary to add password to shell
commands. Now the file attribute can be used for session parameters and
for the bind-auth parameter of tpm2_startauthsession.
So also in these cases the hidden input from stdin can be used.
Two integration tests were extended to test the file attribute.

Examples:

  tpm2_startauthsession --policy-session -S session2.ctx \\
                        --bind-context=prim.ctx --bind-auth=file:-

  tpm2_sign -c key.ctx -p session:session2.ctx+file:- -g sha256 \\
            -o sig -d /tmp/digest

Fixes: #3191

Signed-off-by: Juergen Repp <juergen_repp@web.de>