Add support for parsing forward seal TPM values
Forward sealing is sealing TPM data to precomputed values instead of the
current PCR values.  This can be useful for sealing to new values during
an upgrade process.

Expand the pcr_parse_selections() to save forward seal values to the
new, optional struct tpm2_forwards argument.  This stores all the
provided digest values so they can be later passed to
tpm2_build_policy_pcr().  Forward seal values are only allowed to be
parsed when the argument is not NULL.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>