Use forward seal values in creating policies.
Forward sealing is sealing TPM data to precomputed values instead of the
current PCR values.  This can be useful for sealing to new values during
an upgrade process.

tpm2_policy_build_pcr() is expanded to accept a tpm2_forwards and modify
the digests before passing them on to tpm2_openssl_hash_pcr_values().

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>