files: fix upper bound on unique data size
Originally the unique data was only loadable from a file, and
it had to contain the full TPMU_PUBLIC_ID contents.

When the code was refactored to allow loading from stdin, the
data format was changed:

  commit 8540e7f0edcfcccf6097f0e453fd6c08657a0577
  Author: Imran Desai <imran.desai@intel.com>
  Date:   Tue Oct 13 16:21:08 2020 -0700

    tpm2_createprimary: Add option to input unique from stdin

    Fixes #2091

    Signed-off-by: Imran Desai <imran.desai@intel.com>

In this change the size of the buffer used was altered from
sizeof(TPMU_PUBLIC_ID) to TPM2_MAX_RSA_KEY_BYTES. This is
incorrect as sizeof(TPMU_PUBLIC_ID) is 2 bytes larger than
TPM2_MAX_RSA_KEY_BYTES, as it has the UINT16 size field,
and when loading from a file we need to include those two
extra bytes.

Thus this change caused a regression in the ability to load
unique data from a file for a 512 byte RSA key, since the
file would be 514 bytes, but the buffer only 512.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>