tpm2_createek: Fix usage of nonce.
The current TCG EK Credential Profile [0] in section
"2.2.1.6 TPMT_PUBLIC Calculation" states:
"If the EK Template is Populated and the EK Nonce is Absent,
the EK Template is used unmodified as the TPMT_PUBLIC."
Currently the command did produce an error if there is
a template but no nonce.
Now the template is used but it's checked whether the
nonce exists. Fixes #3278.

Signed-off-by: Juergen Repp <juergen_repp@web.de>