Save Context: Add the possibility for autoflush.
If the environment variable TPM2TOOLS_AUTOFLUSH exists transient
objects will be removed after they were saved and stored to disk.
Also a transient parent will be removed if a context file for
this parent was used by this command.
For the commands which will check the autoflush also an option
-R is added to enable the autoflush independent from the environment
variable.
-R was added to several commands in one integration test.
Addresses: #1511

Signed-off-by: Juergen Repp <juergen_repp@web.de>