tpm2_sessionconfig fix usage of --disable-continuesession
If continue session was disabled a error did occur in the function for
restoring the session context.
Now after usage of an session with continue session disabled the
context will not be saved and the session context file will be
deleted.
In one integration test continue session is now disabled and the
flush for this session is removed.

Fixes: #3295

Signed-off-by: Juergen Repp <juergen_repp@web.de>